pyproject.toml
[project]
name = "webooks-back-joongobooks-v2"
version = "2.0.0"
description = ""
authors = [
    {name = "simseulnyang"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-allauth (>=65.11.2,<66.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=46.0.1,<47.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "channels (>=4.3.2,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
